refactor: base DownloadController tests class, compatibility controller tests
